move shifted_color_map from helpers to plot_helpers and add tests
